@@ -135,7 +135,7 @@ namespace iSukces.Code.AutoCode
             var typeNamespace = type.Namespace ?? "";
             var canCut        = container?.IsKnownNamespace(typeNamespace) ?? false;
             var result = canCut 
-                ? new CsType(mainPart[(typeNamespace.Length + 1)..]) 
+                ? new CsType(mainPart.Substring(typeNamespace.Length + 1)) 
                 : new CsType(mainPart);
             result.GenericParamaters = generics;
             return result;
